+++ source/common/modules/window-register/application-menu-helper.ts
@@ -40,6 +40,7 @@ export interface SubmenuItem {
 export interface NormalItem {
   id: string
   label: string
+  sublabel?: string
   accelerator?: string
   type: 'normal'
   enabled?: boolean